remove multi-release jar contents
