+++ pfl-basic/src/main/java/org/glassfish/pfl/basic/reflection/Bridge.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2018 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2024 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Distribution License v. 1.0, which is available at
@@ -10,24 +10,18 @@
 
 package org.glassfish.pfl.basic.reflection;
 
-import sun.reflect.ReflectionFactory;
-
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.io.ObjectStreamClass;
 import java.io.OptionalDataException;
-import java.io.Serializable;
 import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Constructor;
-import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Field;
 import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.security.AccessController;
 import java.security.Permission;
 import java.security.PrivilegedAction;
-import java.security.PrivilegedExceptionAction;
-import java.util.Objects;
+import java.util.Optional;
+import java.util.stream.Stream;
+
+import sun.reflect.ReflectionFactory;
 
 /**
  * This class provides the methods for fundamental JVM operations
@@ -64,71 +58,12 @@ public final class Bridge extends BridgeBase {
     private static final Permission GET_BRIDGE_PERMISSION = new BridgePermission("getBridge");
     private static Bridge bridge = null;
 
-    // latestUserDefinedLoader() is a private static method
-    // in ObjectInputStream in JDK 1.3 through 1.5.
-    // We use reflection in a doPrivileged block to get a
-    // Method reference and make it accessible.
-    private final Method latestUserDefinedLoaderMethod;
-
-    // Since java.io.OptionalDataException's constructors are
-    // package private, but we need to throw it in some special
-    // cases, we try to do it by reflection.
-    private final Constructor<OptionalDataException> optionalDataExceptionConstructor;
-
     private final ReflectionFactory reflectionFactory;
 
-    private Method getLatestUserDefinedLoaderMethod() {
-        return AccessController.doPrivileged(
-                new PrivilegedAction<Method>() {
-                    @SuppressWarnings("unchecked")
-                    public Method run() {
-                        Method result;
-
-                        try {
-                            Class io = ObjectInputStream.class;
-                            result = io.getDeclaredMethod("latestUserDefinedLoader");
-                            result.setAccessible(true);
-                        } catch (NoSuchMethodException nsme) {
-                            throw new Error("java.io.ObjectInputStream latestUserDefinedLoader " + nsme, nsme);
-                        }
-
-                        return result;
-                    }
-                }
-        );
-    }
-
-    // Grab the OptionalDataException boolean ctor and make it accessible.
-    @SuppressWarnings("unchecked")
-    private Constructor<OptionalDataException> getOptDataExceptionCtor() {
-        try {
-            Constructor result = AccessController.doPrivileged(
-                new PrivilegedExceptionAction<Constructor>() {
-                    public Constructor run()
-                        throws NoSuchMethodException, SecurityException {
-                        Constructor constructor = OptionalDataException.class.getDeclaredConstructor(Boolean.TYPE);
-                        constructor.setAccessible(true);
-                        return constructor;
-                    }
-                }
-            );
-
-            if (result == null) {
-                throw new Error("Unable to find OptionalDataException constructor");
-            }
-
-            return (Constructor<OptionalDataException>) result;
-
-        } catch (Exception ex) {
-            throw new Error("Unable to find OptionalDataException constructor");
-        }
-    }
 
 
     private Bridge() {
-        latestUserDefinedLoaderMethod = getLatestUserDefinedLoaderMethod();
         reflectionFactory = ReflectionFactory.getReflectionFactory();
-        optionalDataExceptionConstructor = getOptDataExceptionCtor();
     }
 
     /**
@@ -157,196 +92,106 @@ public final class Bridge extends BridgeBase {
         return bridge;
     }
 
+
+    private final StackWalker stackWalker = StackWalker.getInstance(StackWalker.Option.RETAIN_CLASS_REFERENCE);
+
+    // New implementation for Java 9, supplied by Alan Bateman
     @Override
     public final ClassLoader getLatestUserDefinedLoader() {
-        try {
-            return (ClassLoader) latestUserDefinedLoaderMethod.invoke(null);
-        } catch (InvocationTargetException | IllegalAccessException ite) {
-            throw new Error(getClass().getName() + ".latestUserDefinedLoader: " + ite, ite);
-        }
+        // requires getClassLoader permission => needs doPrivileged.
+        PrivilegedAction<ClassLoader> pa = () ->
+                stackWalker.walk(this::getLatestUserDefinedLoaderFrame)
+                        .map(sf -> sf.getDeclaringClass().getClassLoader())
+                        .orElseGet(ClassLoader::getPlatformClassLoader);
+        return AccessController.doPrivileged(pa);
+    }
+
+    private Optional<StackWalker.StackFrame> getLatestUserDefinedLoaderFrame(Stream<StackWalker.StackFrame> stream) {
+        return stream.filter(this::isUserLoader).findFirst();
+    }
+
+    private boolean isUserLoader(StackWalker.StackFrame sf) {
+        ClassLoader cl = sf.getDeclaringClass().getClassLoader();
+        if (cl == null) return false;
+
+        ClassLoader platformClassLoader = ClassLoader.getPlatformClassLoader();
+        while (platformClassLoader != null && cl != platformClassLoader) platformClassLoader = platformClassLoader.getParent();
+        return cl != platformClassLoader;
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     public final <T> Constructor<T> newConstructorForExternalization(Class<T> cl) {
-        try {
-            Constructor<T> cons = cl.getDeclaredConstructor();
-            cons.setAccessible(true);
-            return isPublic(cons) ? cons : null;
-        } catch (NoSuchMethodException ex) {
-            return null;
-        }
-    }
-
-    private static boolean isPublic(Constructor<?> cons) {
-        return (cons.getModifiers() & Modifier.PUBLIC) != 0;
+        return (Constructor<T>) reflectionFactory.newConstructorForExternalization( cl );
     }
 
     @Override
     @SuppressWarnings("unchecked")
     public final <T> Constructor<T> newConstructorForSerialization(Class<T> aClass, Constructor<?> cons) {
-        Constructor newConstructor = reflectionFactory.newConstructorForSerialization(aClass, cons);
-        newConstructor.setAccessible(true);
-        return (Constructor<T>) newConstructor;
+        return (Constructor<T>) reflectionFactory.newConstructorForSerialization(aClass, cons);
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     public <T> Constructor<T> newConstructorForSerialization(Class<T> aClass) {
-        Class<?> baseClass = getNearestNonSerializableBaseClass(aClass);
-        if (baseClass == null) return null;
-
-        try {
-            Constructor<?> cons = baseClass.getDeclaredConstructor();
-            if (isPrivate(cons) || !isAccessibleFromSubclass(cons, aClass, baseClass)) return null;
-
-            return newConstructorForSerialization(aClass, cons);
-        } catch (NoSuchMethodException ex) {
-            return null;
-        }
-    }
-
-    private static <T> Class<?> getNearestNonSerializableBaseClass(Class<T> clazz) {
-        Class<?> baseClass = clazz;
-
-        while (Serializable.class.isAssignableFrom(baseClass))
-            if ((baseClass = baseClass.getSuperclass()) == null) return null;
-
-        return baseClass;
-    }
-
-    private static boolean isAccessibleFromSubclass(Constructor<?> constructor, Class<?> clazz, Class<?> baseClass) {
-        return isPublicOrProtected(constructor) || inSamePackage(clazz, baseClass);
-    }
-
-    private static boolean inSamePackage(Class<?> clazz, Class<?> baseClass) {
-        return Objects.equals(clazz.getPackage(), baseClass.getPackage());
-    }
-
-    private static boolean isPublicOrProtected(Constructor<?> constructor) {
-        return (constructor.getModifiers() & (Modifier.PUBLIC | Modifier.PROTECTED)) != 0;
-    }
-
-    private static boolean isPrivate(Constructor<?> cons) {
-        return (cons.getModifiers() & Modifier.PRIVATE) != 0;
-    }
-
-    private static Method hasStaticInitializerMethod = null;
-
-    @Override
-    public boolean hasStaticInitializerForSerialization(Class<?> cl) {
-        try {
-            return (Boolean) getHasStaticInitializerMethod().invoke(null, cl);
-        } catch (Exception ex) {
-            throw new Error("Cannot invoke 'hasStaticInitializer' method on " + ObjectStreamClass.class.getName());
-        }
-    }
-
-    private static Method getHasStaticInitializerMethod() throws NoSuchMethodException {
-        if (hasStaticInitializerMethod == null) {
-            hasStaticInitializerMethod = ObjectStreamClass.class.getDeclaredMethod("hasStaticInitializer", Class.class);
-            hasStaticInitializerMethod.setAccessible(true);
-        }
-
-        return hasStaticInitializerMethod;
-    }
-
-    @Override
-    public MethodHandle writeObjectForSerialization(Class<?> cl) {
-        return toMethodHandle(getPrivateMethod(cl, "writeObject", Void.TYPE, ObjectOutputStream.class));
-    }
-
-    private static MethodHandle toMethodHandle(Method method) {
-        try {
-            if (method == null) return null;
-            method.setAccessible(true);
-            MethodHandle methodHandle = MethodHandles.lookup().unreflect(method);
-            method.setAccessible(false);
-            return methodHandle;
-        } catch (SecurityException | IllegalAccessException e) {
-            return null;
-        }
-    }
-
-    private static Method getPrivateMethod(Class<?> cl, String name, Class<?> returnType, Class<?>... argTypes ) {
-        try {
-            Method method = cl.getDeclaredMethod(name, argTypes);
-            return ((method.getReturnType() == returnType) && isPrivate(method) && !isStatic(method)) ? method : null;
-        } catch (NoSuchMethodException ex) {
-            return null;
-        }
-    }
-
-    private static boolean isStatic(Method method) {
-        return Modifier.isStatic(method.getModifiers());
-    }
-
-    private static boolean isPrivate(Method method) {
-        return Modifier.isPrivate(method.getModifiers());
-    }
-
-    @Override
-    public MethodHandle readObjectForSerialization(Class<?> cl) {
-        return toMethodHandle(getPrivateMethod(cl, "readObject", Void.TYPE, ObjectInputStream.class));
-    }
-
-    @Override
-    public MethodHandle readResolveForSerialization(Class<?> cl) {
-        return toMethodHandle(getInheritableMethod(cl, "readResolve", Object.class));
-    }
-
-    private static Method getInheritableMethod(Class<?> cl, String name, Class<?> returnType, Class<?>... argTypes ) {
-        Method method = getMatchingMethod(cl, name, returnType, argTypes);
-
-        return (method != null && isMethodInheritableBy(cl, method)) ? method : null;
-    }
-
-    private static Method getMatchingMethod(Class<?> cl, String name, Class<?> returnType, Class<?>[] argTypes) {
-        Class<?> aClass = cl;
-        while (aClass != null) {
-            try {
-                Method method = aClass.getDeclaredMethod(name, argTypes);
-                return method.getReturnType() == returnType ? method : null;
-            } catch (NoSuchMethodException ex) {
-                aClass = aClass.getSuperclass();
-            }
-        }
-        return null;
-    }
-
-    private static boolean isMethodInheritableBy(Class<?> callingClass, Method method) {
-        Class<?> baseClass = method.getDeclaringClass();
-
-        int mods = method.getModifiers();
-        if ((mods & (Modifier.STATIC | Modifier.ABSTRACT)) != 0) {
-            return false;
-        } else if ((mods & (Modifier.PUBLIC | Modifier.PROTECTED)) != 0) {
-            return true;
-        } else if ((mods & Modifier.PRIVATE) != 0) {
-            return (callingClass == baseClass);
-        } else {
-            return packageEquals(callingClass, baseClass);
-        }
+        return (Constructor<T>) reflectionFactory.newConstructorForSerialization( aClass );
     }
 
     /**
-     * Returns true if classes are defined in the same package, false
-     * Copied from the Merlin java.io.ObjectStreamClass.
+     * Returns true if the given class defines a static initializer method,
+     * false otherwise.
      */
-    private static boolean packageEquals(Class<?> cl1, Class<?> cl2) {
-        Package pkg1 = cl1.getPackage(), pkg2 = cl2.getPackage();
-        return ((pkg1 == pkg2) || ((pkg1 != null) && (pkg1.equals(pkg2))));
+    @Override
+    public final boolean hasStaticInitializerForSerialization(Class<?> cl) {
+        return reflectionFactory.hasStaticInitializerForSerialization(cl);
     }
 
     @Override
-    public MethodHandle writeReplaceForSerialization(Class<?> cl) {
-        return toMethodHandle(getInheritableMethod(cl, "writeReplace", Object.class));
+    public final MethodHandle writeObjectForSerialization(Class<?> cl) {
+        return reflectionFactory.writeObjectForSerialization(cl);
     }
 
     @Override
-    public OptionalDataException newOptionalDataExceptionForSerialization(boolean endOfData) {
-        try {
-            return optionalDataExceptionConstructor.newInstance(endOfData);
-        } catch (InstantiationException | IllegalAccessException | InvocationTargetException e) {
-            throw new Error("Unable to create OptionalDataException");
-        }
+    public final MethodHandle readObjectForSerialization(Class<?> cl) {
+        return reflectionFactory.readObjectForSerialization(cl);
+    }
+
+    public final MethodHandle readObjectNoDataForSerialization(Class<?> cl) {
+        return reflectionFactory.readObjectNoDataForSerialization(cl);
+    }
+
+    @Override
+    public final MethodHandle readResolveForSerialization(Class<?> cl) {
+        return reflectionFactory.readResolveForSerialization(cl);
+    }
+
+    @Override
+    public final MethodHandle writeReplaceForSerialization(Class<?> cl) {
+        return reflectionFactory.writeReplaceForSerialization(cl);
+    }
+
+    @Override
+    public final OptionalDataException newOptionalDataExceptionForSerialization(boolean bool) {
+        return reflectionFactory.newOptionalDataExceptionForSerialization(bool);
+    }
+
+    @Override
+    public Field toAccessibleField(Field field, Class<?> callingClass) {
+        return isClassOpenToModule(field.getDeclaringClass(), callingClass.getModule())
+              ? super.toAccessibleField(field, callingClass)
+              : null;
+    }
+
+    private boolean isClassOpenToModule(Class<?> candidateClass, Module callingModule) {
+        return callingModule.isNamed()
+              ? candidateClass.getModule().isOpen(candidateClass.getPackageName(), callingModule)
+              : candidateClass.getModule().isOpen(candidateClass.getPackageName());
+    }
+
+    @Override
+    public Method toAccessibleMethod(Method method, Class<?> callingClass) {
+        return isClassOpenToModule(method.getDeclaringClass(), callingClass.getModule())
+              ? super.toAccessibleMethod(method, callingClass)
+              : null;
     }
 }
